replaced floats with double, trying to fix warnings in win32.
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@1280 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/components/laser2d/sickcarmen/sickcarmendriver.cpp
@@ -60,8 +60,8 @@ SickCarmenDriver::setupParams( double maxRange, int numberOfReturns, int baudrat
     // first convert from generic configs to SICK-specific ones
     //
 
-    float rangeResolution = sickutil::maxRange2RangeResolution( maxRange );
-    float angleIncrement = sickutil::numberOfReturns2AngleIncrement( numberOfReturns );
+    double rangeResolution = sickutil::maxRange2RangeResolution( maxRange );
+    double angleIncrement = sickutil::numberOfReturns2AngleIncrement( numberOfReturns );
     
     stringstream ss;
     ss<<"Validating SICK config: rangeres="<<rangeResolution<<" angleincrem="<<angleIncrement<<" baud="<<baudrate;
